F #4937: reorganize fsck code
This contains all these changes

* Use IMAGE_STATES array from OCA
* Including OpenNebula breaks fsck
* Extract check_pool_control and check_users_groups
* Extract check_clusters
* Extract check_datastores
* Move schema definitions to a hash
* Move counter initialization to its own methods
* Add comments to sections

To see them use:

    grep -E '(DATA|FIX|TODO):' fsck.rb

* More fsck comments
* Move image checks to its own file
* Move marketplace/app checks to it's own file
* Take out do_ prefix from functions
* Move pool_control checks
* Move user and group checks
* Split check_clusters to external functions

There are still functions that regenerate the whole table. This should
be fixed later. Look for functions called check_fix_*.

Look also for use of REXML instead of Nokogiri. This can be found
searching for "Document.new".

Some searches are done using plain SQL statements. These can be
changed to sequel functions.

* Split check_datastores
* Move VM checks
* Move cluster_vnc_bitmap checks
* Move history check
* Move vrouter check
* Move host checks
* Move network checks
* Move quota checks
* Move template checks
* Reorganize quota check code
* Calculate quota for datastores with target SELF
* Add fsck directory to install.sh
* Bug in fsck quota code
