+++ install.sh
@@ -1157,7 +1157,8 @@ ONEDB_FILES="src/onedb/fsck.rb \
             src/onedb/import_slave.rb \
             src/onedb/onedb.rb \
             src/onedb/onedb_backend.rb \
-            src/onedb/sqlite2mysql.rb"
+            src/onedb/sqlite2mysql.rb \
+            src/onedb/fsck"
 
 ONEDB_SHARED_MIGRATOR_FILES="src/onedb/shared/2.0_to_2.9.80.rb \
                              src/onedb/shared/2.9.80_to_2.9.85.rb \
